iter2: Move handler
