+++ cmd/gophermart/main.go
@@ -9,7 +9,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/bjlag/go-loyalty/internal/api/handler/balance"
+	"github.com/bjlag/go-loyalty/internal/api/handler/balance/get"
 	"github.com/bjlag/go-loyalty/internal/api/handler/order/list"
 	"github.com/bjlag/go-loyalty/internal/api/handler/order/upload"
 	"github.com/bjlag/go-loyalty/internal/api/handler/user/login"
@@ -87,7 +87,7 @@ func main() {
 		withAPIHandler(http.MethodPost, "/api/user/login", login.NewHandler(usecaseLogin, log).Handle),
 		withAPIHandler(http.MethodPost, "/api/user/orders", upload.NewHandler(usecaseCreateAccrual, log).Handle, middleware.CheckAuth(jwtBuilder, log)),
 		withAPIHandler(http.MethodGet, "/api/user/orders", list.NewHandler(accrualRepo, log).Handle, middleware.CheckAuth(jwtBuilder, log)),
-		withAPIHandler(http.MethodGet, "/api/user/balance", balance.NewHandler(accountRepo, log).Handle, middleware.CheckAuth(jwtBuilder, log)),
+		withAPIHandler(http.MethodGet, "/api/user/balance", get.NewHandler(accountRepo, log).Handle, middleware.CheckAuth(jwtBuilder, log)),
 	)
 
 	if err := app.run(ctx); err != nil {
